+++ src/netaddr_error.rs
@@ -1,0 +1,21 @@
+use core::fmt::{Display, Error, Formatter};
+use std::net::AddrParseError;
+
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum NetAddrError {
+	ParseError(String),
+}
+
+impl Display for NetAddrError {
+	fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+		match self {
+			NetAddrError::ParseError(text) => write!(f, "unable to parse address: {}", text),
+		}
+	}
+}
+
+impl From<AddrParseError> for NetAddrError {
+	fn from(other: AddrParseError) -> Self {
+		NetAddrError::ParseError(other.to_string())
+	}
+}
